NETOBSERV-184: decouple components into a pipeline
As a previous step for NETOBSERV-184 (Split and Integrate flowlogs2metrics in
kube-enricher), this PR provides an asynchronous pipeline and wraps the current
components into the following roles:

* Ingester: a wrapped version of Goflow2
* Transformer: the Pod & Service Enricher
* Submitter: the Loki processor

This is only a previous step of the above task to avoid having too large PRs.
The following PRs will try to simplify and decouple even more some existing
components (e.g. Goflow-Kube), remove duplicities. After that, we will
make the Pipeline object to evolve to accomodate the new functionalities
from flowlogs2metrics